No deberia haber comiteado esto

+++ g2g/cuda/cuda_extra.h
@@ -292,13 +292,13 @@ inline void cudaAssertNoError(const char* msg = NULL) {
 #define cudaAssertNoError(s)
 #endif
 
-inline void cudaGetMemoryInfo(ulong& free, ulong& total) {
+inline void cudaGetMemoryInfo(uint& free, uint& total) {
 	if (cuMemGetInfo(&free, &total) != CUDA_SUCCESS)
     throw std::runtime_error("cuMemGetInfo failed");
 }
 
 inline void cudaPrintMemoryInfo(void) {
-  ulong free = 0, total = 0;
+  uint free = 0, total = 0;
   cudaGetMemoryInfo(free, total);
 	std::cout << "mem_used: " << (total - free) / (1024.0 * 1024.0) << "MB | mem_perc: " << ((double)(total - free) / (double)total) * 100.0 << "%" << std::endl;
 }
